Move Kfp-specific properties to their own file
