+++ elyra/pipeline/properties.py
@@ -20,7 +20,6 @@ from abc import abstractmethod
 from enum import Enum
 from importlib import import_module
 import json
-from keyword import iskeyword
 import re
 from typing import Any
 from typing import Dict
@@ -63,34 +62,38 @@ class PropertyInputType:
         "bool": {"json_type": "boolean", "default_value": False, "placeholder": " "},
         "int": {"json_type": "integer"},
         "float": {"json_type": "number"},
-        "list": {"json_type": "array", "default_value": []},
-        "dict": {"json_type": "dict", "default_value": {}},
+        "list": {"json_type": "string", "default_value": "[]"},
+        "dict": {"json_type": "string", "default_value": "{}"},
     }
 
     def __init__(
         self,
-        base_input_type: str,
+        base_type: str,
         default_value: Optional[Any] = None,
         placeholder: Optional[Any] = None,
         enum: Optional[List[Any]] = None,
-        runtime_specific_types: Optional[Dict[str, Dict]] = None,
+        **kwargs,
     ):
         # Get type info from passed in type dict, if available; default to standard python types
-        allowed_input_types = runtime_specific_types or self._python_input_types
-        if base_input_type not in allowed_input_types:
+        allowed_input_types = kwargs.get("allowed_input_types")
+        if not allowed_input_types:
+            allowed_input_types = self._python_input_types
+
+        if base_type not in allowed_input_types:
             raise ValueError(
-                f"Invalid property type '{base_input_type}': valid types are f{list(allowed_input_types.keys())}."
+                f"Invalid property type '{base_type}': valid types are f{list(allowed_input_types.keys())}."
             )
-        type_defaults = allowed_input_types[base_input_type]
+        type_defaults = allowed_input_types[base_type]
 
         # Set types needed for schema
-        self.base_input_type = base_input_type
+        self.base_type = base_type
         self.json_data_type = type_defaults.get("json_type")
 
         # Set extra schema details
         self.default_value = default_value if default_value is not None else type_defaults.get("default_value")
         self.placeholder = placeholder if placeholder is not None else type_defaults.get("placeholder")
         self.enum = enum
+        self.type_title = type_defaults.get("type_title")
 
 
 class PropertyAttribute:
@@ -274,16 +277,17 @@ class ElyraProperty(ABC):
                 # Render a oneOf block
                 properties[attr.id]["oneOf"] = []
                 for allowed_type in attr.allowed_input_types:
-                    display_title = f"Parameter value type: {allowed_type.base_input_type}"
                     schema_obj = {
                         "type": "object",
-                        "title": display_title,
+                        "title": f"Parameter type: {allowed_type.type_title or allowed_type.base_type}",
                         "properties": {
-                            "type": {"type": "string", "default": allowed_type.base_input_type},
-                            "value": {"type": allowed_type.json_data_type, "title": allowed_type.base_input_type},
+                            "type": {"type": "string", "default": allowed_type.base_type},
+                            "value": {"type": allowed_type.json_data_type},
                         },
                         "uihints": {"type": {"ui:widget": "hidden"}},
                     }
+                    if allowed_type.json_data_type == "boolean":
+                        schema_obj["properties"]["value"]["title"] = "Check for true"
 
                     if allowed_type.default_value is not None:
                         schema_obj["properties"]["value"]["default"] = allowed_type.default_value
@@ -404,14 +408,14 @@ class CustomSharedMemorySize(ElyraProperty):
         PropertyAttribute(
             attribute_id="size",
             display_name="Memory Size (GB)",
-            allowed_input_types=[PropertyInputType(base_input_type="int", placeholder=0)],
+            allowed_input_types=[PropertyInputType(base_type="int", placeholder=0)],
             hidden=False,
             required=False,
         ),
         PropertyAttribute(
             attribute_id="units",
             display_name="Units",
-            allowed_input_types=[PropertyInputType(base_input_type="str")],
+            allowed_input_types=[PropertyInputType(base_type="str")],
             hidden=True,
             required=False,
         ),
@@ -516,7 +520,7 @@ class EnvironmentVariable(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="env_var",
             display_name="Environment Variable",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="ENV_VAR")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="ENV_VAR")],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -524,7 +528,7 @@ class EnvironmentVariable(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="value",
             display_name="Value",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="value")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="value")],
             hidden=False,
             required=False,
             use_in_key=False,
@@ -583,7 +587,7 @@ class KubernetesSecret(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="env_var",
             display_name="Environment Variable",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="ENV_VAR")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="ENV_VAR")],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -591,7 +595,7 @@ class KubernetesSecret(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="name",
             display_name="Secret Name",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="secret-name")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="secret-name")],
             hidden=False,
             required=True,
             use_in_key=False,
@@ -599,7 +603,7 @@ class KubernetesSecret(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="key",
             display_name="Secret Key",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="secret-key")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="secret-key")],
             hidden=False,
             required=True,
             use_in_key=False,
@@ -650,7 +654,7 @@ class VolumeMount(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="path",
             display_name="Mount Path",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="/mount/path")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="/mount/path")],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -658,7 +662,7 @@ class VolumeMount(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="pvc_name",
             display_name="Persistent Volume Claim Name",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="pvc-name")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="pvc-name")],
             hidden=False,
             required=True,
             use_in_key=False,
@@ -666,7 +670,7 @@ class VolumeMount(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="sub_path",
             display_name="Sub Path",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="relative/path/within/volume")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="relative/path/within/volume")],
             hidden=False,
             required=False,
             use_in_key=False,
@@ -674,7 +678,7 @@ class VolumeMount(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="read_only",
             display_name="Mount volume read-only",
-            allowed_input_types=[PropertyInputType(base_input_type="bool", default_value=False, placeholder=None)],
+            allowed_input_types=[PropertyInputType(base_type="bool", default_value=False, placeholder=None)],
             hidden=False,
             required=False,
             use_in_key=False,
@@ -732,7 +736,7 @@ class KubernetesAnnotation(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="key",
             display_name="Key",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="annotation_key")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="annotation_key")],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -740,7 +744,7 @@ class KubernetesAnnotation(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="value",
             display_name="Value",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="annotation_value")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="annotation_value")],
             hidden=False,
             required=False,
             use_in_key=False,
@@ -791,7 +795,7 @@ class KubernetesLabel(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="key",
             display_name="Key",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="label_key")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="label_key")],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -799,7 +803,7 @@ class KubernetesLabel(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="value",
             display_name="Value",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="label_value")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="label_value")],
             hidden=False,
             required=False,
             use_in_key=False,
@@ -848,7 +852,7 @@ class KubernetesToleration(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="key",
             display_name="Key",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="key")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="key")],
             hidden=False,
             required=False,
             use_in_key=True,
@@ -856,9 +860,7 @@ class KubernetesToleration(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="operator",
             display_name="Operator",
-            allowed_input_types=[
-                PropertyInputType(base_input_type="str", default_value="Equal", enum=["Equal", "Exists"])
-            ],
+            allowed_input_types=[PropertyInputType(base_type="str", default_value="Equal", enum=["Equal", "Exists"])],
             hidden=False,
             required=True,
             use_in_key=True,
@@ -866,7 +868,7 @@ class KubernetesToleration(ElyraPropertyListItem):
         ListItemPropertyAttribute(
             attribute_id="value",
             display_name="Value",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="value")],
+            allowed_input_types=[PropertyInputType(base_type="str", placeholder="value")],
             hidden=False,
             required=False,
             use_in_key=True,
@@ -876,7 +878,7 @@ class KubernetesToleration(ElyraPropertyListItem):
             display_name="Effect",
             allowed_input_types=[
                 PropertyInputType(
-                    base_input_type="str",
+                    base_type="str",
                     placeholder="NoSchedule",
                     enum=["", "NoExecute", "NoSchedule", "PreferNoSchedule"],
                 )
@@ -1233,7 +1235,7 @@ class PipelineParameter(ElyraPropertyListItem):
         self.default_value = default_value.get("value") if isinstance(default_value, dict) else None
 
         # Assign value, if set; use default_value if not
-        self.value = value.get("value") if isinstance(value, dict) else self.default_value
+        self.value = value if value is not None else self.default_value  # TODO may change with pipeline JSON
 
     def get_value_for_dict_entry(self) -> Union[str, Dict[str, Any]]:
         """
@@ -1256,116 +1258,3 @@ class PipelineParameter(ElyraPropertyListItem):
 
     def add_to_execution_object(self, runtime_processor: RuntimePipelineProcessor, execution_object: Any, **kwargs):
         raise RuntimeError("Method should not be invoked.")
-
-
-class KfpPropertyInputType(PropertyInputType):
-    """
-    An object representing a single allowed input type for a PropertyAttribute
-    object for KUBEFLOW_PIPELINES runtime processors.
-    """
-
-    _allowed_input_types = {
-        "String": {"python_type": "str", "json_type": "string", "default_value": ""},
-        "Bool": {"python_type": "bool", "json_type": "boolean", "default_value": False, "placeholder": " "},
-        "Integer": {"python_type": "int", "json_type": "integer"},
-        "Float": {"python_type": "float", "json_type": "number"},
-        "List": {"python_type": "list", "json_type": "array", "default_value": []},
-        "Dict": {"python_type": "dict", "json_type": "object", "default_value": {}},
-    }
-
-    def __init__(
-        self,
-        base_input_type: str,
-        default_value: Optional[Any] = None,
-        placeholder: Optional[Any] = None,
-        enum: Optional[List[Any]] = None,
-    ):
-        super().__init__(
-            base_input_type=base_input_type,
-            default_value=default_value,
-            placeholder=placeholder,
-            enum=enum,
-            runtime_specific_types=self._allowed_input_types,
-        )
-
-        self.python_type = self._allowed_input_types[base_input_type].get("python_type") or f'"{base_input_type}"'
-
-
-class KfpPipelineParameter(PipelineParameter):
-    """An ElyraProperty representing a single pipeline parameter for the Kubeflow Pipelines runtime"""
-
-    property_id = "KFP_PIPELINE_PARAMETERS"
-    property_attributes = [
-        ListItemPropertyAttribute(
-            attribute_id="name",
-            display_name="Parameter Name",
-            allowed_input_types=[PropertyInputType(base_input_type="str", placeholder="param_1")],
-            hidden=False,
-            required=True,
-            use_in_key=True,
-        ),
-        ListItemPropertyAttribute(
-            attribute_id="default_value",
-            display_name="Default Value",
-            allowed_input_types=[
-                KfpPropertyInputType(base_input_type="String", placeholder="default_val"),
-                KfpPropertyInputType(base_input_type="Integer"),
-                KfpPropertyInputType(base_input_type="Float"),
-                KfpPropertyInputType(base_input_type="Bool"),
-                KfpPropertyInputType(base_input_type="List"),
-                KfpPropertyInputType(base_input_type="Dict"),
-            ],
-            hidden=False,
-            required=False,
-            use_in_key=False,
-        ),
-        ListItemPropertyAttribute(
-            attribute_id="value",
-            display_name="Value",
-            allowed_input_types=[
-                KfpPropertyInputType(base_input_type="String"),
-                KfpPropertyInputType(base_input_type="Integer"),
-                KfpPropertyInputType(base_input_type="Float"),
-                KfpPropertyInputType(base_input_type="Bool"),
-                KfpPropertyInputType(base_input_type="List"),
-                KfpPropertyInputType(base_input_type="Dict"),
-            ],
-            hidden=True,
-            required=False,
-            use_in_key=False,
-        ),
-        ListItemPropertyAttribute(
-            attribute_id="required",
-            display_name="Required",
-            allowed_input_types=[PropertyInputType(base_input_type="bool", placeholder=" ")],
-            hidden=False,
-            required=True,
-            use_in_key=False,
-        ),
-    ]
-
-    def __init__(self, name, value, default_value, required, **kwargs):
-        super().__init__(name=name, value=value, default_value=default_value, required=required)
-        self.input_type = KfpPropertyInputType(base_input_type=value.get("type"))
-        # TODO Coerce number types to ints and floats if needed
-
-    def get_all_validation_errors(self) -> List[str]:
-        """Perform custom validation on an instance."""
-        validation_errors = []
-        if not self.name:
-            validation_errors.append("Required parameter name was not specified.")
-        elif not self.name.isidentifier():
-            # param name is not a valid python variable name
-            validation_errors.append(
-                f"'{self.name}' is not a valid parameter name: name must be a Python variable name."
-            )
-        elif iskeyword(self.name):
-            # param name collides with a python keyword (e.g. class, def, etc.)
-            validation_errors.append(f"'{self.name}' is not a valid parameter name: name cannot be a Python keyword.")
-
-        # If 'required' is True, a value must be provided
-        if self.required and self.value is None or self.value == "":
-            validation_errors.append("Parameter is marked as required but no value has been assigned.")
-
-        # TODO check that input type matches the value
-        return validation_errors
