+++ elyra/templates/kubeflow/v1/generic_component_definition_template.jinja2
@@ -3,7 +3,7 @@ description: Run a Jupyter notebook or Python/R script
 {% if task_parameters %}
 inputs:
 {%- for parameter in task_parameters %}
-- {name: {{ parameter.name }}, type: {{ parameter.input_type.base_input_type }}}
+- {name: {{ parameter.name }}, type: {{ parameter.input_type.component_input_type }}}
 {%- endfor %}
 {% endif %}
 implementation:
